requirements.txt
asgiref==3.5.2
celery>=5.2.7
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-admin-rangefilter==0.8.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-import-export==2.8.0
django-phonenumber-field==6.3.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
phonenumbers==8.12.50
psycopg2-binary==2.9.3
pytz==2022.1
PyYAML==6.0
redis>=4.3.4
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
xlrd==2.0.1
xlwt==1.3.0
